Assinment 2/pyproject.toml
[build-system]
requires = ["hatchling>=1.27.0"]
build-backend = "hatchling.build"

[project]
name = "graph-cut-segmentation-assignment"
version = "0.1.0"
description = "Graph cut image segmentation pipeline for CV Assignment 2."
readme = "README.md"
requires-python = ">=3.11,<3.12"
dependencies = [
  "imageio>=2.37.3",
  "matplotlib>=3.10.8",
  "numpy>=2.4.4",
  "opencv-python>=4.13.0.92",
  "PyMaxflow>=1.3.2",
  "PyYAML>=6.0.3",
  "scikit-image>=0.26.0",
  "scipy>=1.17.1",
]

[dependency-groups]
dev = [
  "pytest>=9.0.2",
]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]

[tool.hatch.build.targets.wheel]
packages = ["src"]
